// ShoppingMall-Solo/src/modules/apidata.interface.ts
interface ApiDataInterFace {
  brand: string;
  category: string;
  description: string;
  discountPercentage: number;
  id: number;
  images: string[];
  price: number;
  rating: number;
  stock: number;
  thumbnail: string;
  title: string;
  bookmark?: boolean;
}

export default ApiDataInterFace;